+++ R/afltables_basic.R
@@ -86,6 +86,7 @@ get_match_results <- function() {
   match_data <- match_data %>%
     dplyr::group_by(.data$Game) %>%
     dplyr::mutate_at(c("Home.Team", "Away.Team"), replace_teams) %>%
+    dplyr::mutate(Venue = replace_venues(Venue)) %>%
     dplyr::ungroup()
 
 
@@ -133,5 +134,41 @@ replace_teams <- function(team) {
     team == "Cats" ~ "Geelong",
     team == "Hawks" ~ "Hawthorn",
     TRUE ~ team
+  )
+}
+
+#' Internal function to ensure venue names match between different sources
+#' and also name changes across time. This gets applied to any web scraper,
+#' transforming all of them to AFL Tables naming conventions.
+#' @param venue Venue name
+#' @export
+replace_venues <- function(venue) {
+  # Internal function
+  dplyr::case_when(
+    venue == "AAMI Stadium" ~ "Football Park",
+    venue == "ANZ Stadium" ~ "Stadium Australia",
+    venue == "UTAS Stadium" ~ "York Park",
+    venue == "Blacktown International" ~ "Blacktown",
+    venue == "Blundstone Arena" ~ "Bellerive Oval",
+    venue == "Domain Stadium" ~ "Subiaco",
+    venue == "Etihad Stadium" ~ "Docklands",
+    venue == "Marvel Stadium" ~ "Docklands",
+    venue == "GMHBA Stadium" ~ "Kardinia Park",
+    venue == "MCG" ~ "M.C.G.",
+    venue == "Mars Stadium" ~ "Eureka Stadium",
+    venue == "Metricon Stadium" ~ "Carrara",
+    venue == "Optus Stadium" ~ "Perth Stadium",
+    venue == "SCG" ~ "S.C.G.",
+    venue == "Spotless Stadium" ~ "Sydney Showground",
+    venue == "Showground Stadium" ~ "Sydney Showground",
+    venue == "GIANTS Stadium" ~ "Sydney Showground",
+    venue == "TIO Stadium" ~ "Marrara Oval",
+    venue == "Westpac Stadium" ~ "Wellington",
+    venue == "Canberra Oval" ~ "Manuka Oval",
+    venue == "TIO Traeger Park" ~ "Traeger Park",
+    # Correct spelling is 'Traeger', but footywire.com is spelling it 'Traegar'
+    # in its fixtures, so including both in case they eventually fix the misspelling
+    venue == "TIO Traegar Park" ~ "Traeger Park",
+    TRUE ~ venue
   )
 }